requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
assemblyai==0.35.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.4
entrypoints==0.4
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
numpy==2.2.0
openai==0.28.0
packaging==24.2
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pyepics==3.5.7
pyparsing==3.2.0
pyqtgraph==0.13.7
python-decouple==3.8
QtPy==2.4.2
requests==2.32.3
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
yarl==1.18.3
yt-dlp==2024.12.13